Fixes error

+++ enqueue/enqueueMain.py
@@ -635,6 +635,8 @@ def clearFailed():
         queue = Queue(queue_name, connection=redis_connection)
         for job_id in queue.failed_job_registry.get_job_ids():
             job = queue.fetch_job(job_id)
+            if not job or not job.ended_at:
+                continue
             if job and (datetime.now() - job.ended_at) >= timedelta(hours=hours):
                 job.delete()
                 count += 1
